# account/src/main/resources/swagger.yaml
openapi: 3.0.3
info:
  title: Thadam Bank Project API
  description: API for Account Opening, Deposit and Withdraw
  version: 1.0.0

servers:
  - url: http://localhost:8080/api
    description: Local server

paths:

  /account:
    post:
      tags:
        - Account
      summary: Open a new account
      operationId: openAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountDTO'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDTO'

  /account/deposit:
    post:
      tags:
        - Account
      summary: Deposit amount to account
      operationId: depositAmount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionDTO'
      responses:
        '200':
          description: Amount deposited successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDTO'

  /account/withdraw:
    post:
      tags:
        - Account
      summary: Withdraw amount from account
      operationId: withdrawAmount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionDTO'
      responses:
        '200':
          description: Amount withdrawn successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDTO'

components:
  schemas:

    AccountDTO:
      type: object
      properties:
        accountId:
          type: integer
          example: 101
        accountHolderName:
          type: string
          example: Cathrine
        branch:
          type: string
          example: Madurai
        balance:
          type: integer
          example: 5000
        contact:
          type: string
          example: "9876543210"

    TransactionDTO:
      type: object
      properties:
        accountId:
          type: integer
          example: 101
        amount:
          type: integer
          example: 1000